+++ daml-lf/engine/BUILD.bazel
@@ -141,6 +141,7 @@ da_scala_test(
         "//ledger/ledger-on-memory",
         "//ledger/ledger-on-memory:ledger-on-memory-app",
         "//ledger/ledger-resources",
+        "//ledger/ledger-runner-common",
         "//ledger/participant-integration-api",
         "//ledger/participant-state/kvutils/app",
         "//ledger/test-common:dar-files-1.14-lib",
