+++ daml-script/test/BUILD.bazel
@@ -153,9 +153,8 @@ da_scala_library(
         "//ledger/ledger-on-memory",
         "//ledger/ledger-on-memory:ledger-on-memory-app",
         "//ledger/ledger-resources",
+        "//ledger/ledger-runner-common",
         "//ledger/participant-integration-api",
-        "//ledger/participant-state/kvutils",
-        "//ledger/participant-state/kvutils/app",
         "//ledger/sandbox-common",
         "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
         "//ledger/sandbox-on-x:sandbox-on-x-test-lib",
@@ -211,12 +210,12 @@ da_scala_test_suite(
         "//ledger/ledger-api-common",
         "//ledger/ledger-configuration",
         "//ledger/ledger-resources",
+        "//ledger/ledger-runner-common",
         "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
         "//ledger/participant-state",
         "//ledger/participant-state/kvutils",
-        "//ledger/participant-state/kvutils/app",
         "//ledger/sandbox-common",
         "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
         "//ledger/sandbox-on-x:sandbox-classic-server",
