+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/Config.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.app
+package com.daml.ledger.runner.common
 
 import com.daml.caching
 import com.daml.jwt.JwtVerifierConfigurationCli
