+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/ConfigParseException.scala
@@ -1,10 +1,10 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.app
+package com.daml.ledger.runner.common
 
 import com.daml.resources.ProgramResource.SuppressedStartupException
 
-private[kvutils] final class ConfigParseException
+private[common] final class ConfigParseException
     extends RuntimeException
     with SuppressedStartupException
