+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/DumpIndexMetadata.scala
@@ -1,16 +1,15 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.app
-
-import com.daml.ledger.resources.{Resource, ResourceContext}
-import com.daml.logging.{ContextualizedLogger, LoggingContext}
-import com.daml.logging.LoggingContext.newLoggingContext
-import com.daml.platform.server.api.validation.ErrorFactories
-import com.daml.platform.store.IndexMetadata
+package com.daml.ledger.runner.common
 
 import akka.actor.ActorSystem
 import akka.stream.Materializer
+import com.daml.ledger.resources.{Resource, ResourceContext}
+import com.daml.logging.LoggingContext.newLoggingContext
+import com.daml.logging.{ContextualizedLogger, LoggingContext}
+import com.daml.platform.server.api.validation.ErrorFactories
+import com.daml.platform.store.IndexMetadata
 
 import scala.concurrent.ExecutionContext
 import scala.util.{Failure, Success}
