+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/ParticipantConfig.scala
@@ -1,12 +1,12 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.app
+package com.daml.ledger.runner.common
 
 import java.nio.file.Path
 import java.time.Duration
 
-import com.daml.ledger.participant.state.kvutils.app.ParticipantConfig.DefaultApiServerDatabaseConnectionTimeout
+import com.daml.ledger.runner.common.ParticipantConfig.DefaultApiServerDatabaseConnectionTimeout
 import com.daml.lf.data.Ref
 import com.daml.ports.Port
