+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/ParticipantRunMode.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.app
+package com.daml.ledger.runner.common
 
 sealed abstract class ParticipantRunMode
