+++ ledger/sandbox-on-x/src/classic/scala/com/daml/platform/ledger/ConfigConverter.scala
@@ -5,7 +5,7 @@ package com.daml.ledger.sandbox
 
 import com.daml.caching
 import com.daml.ledger.api.auth.AuthServiceWildcard
-import com.daml.ledger.participant.state.kvutils.app._
+import com.daml.ledger.runner.common._
 import com.daml.lf.language.LanguageVersion
 import com.daml.platform.common.LedgerIdMode
 import com.daml.platform.sandbox.config.SandboxConfig.{DefaultTimeProviderType, EngineMode}
