+++ ledger/sandbox-on-x/src/classic/scala/com/daml/platform/ledger/SandboxServer.scala
@@ -7,7 +7,7 @@ import akka.actor.ActorSystem
 import akka.stream.Materializer
 import com.codahale.metrics.MetricRegistry
 import com.daml.buildinfo.BuildInfo
-import com.daml.ledger.participant.state.kvutils.app.Config
+import com.daml.ledger.runner.common.Config
 import com.daml.ledger.participant.state.v2.WriteService
 import com.daml.ledger.resources
 import com.daml.ledger.resources.{Resource, ResourceContext, ResourceOwner}
