+++ release/artifacts.yaml
@@ -187,6 +187,8 @@
   type: jar-scala
 - target: //ledger/sandbox-on-x:app
   type: jar-deploy
+- target: //ledger/ledger-runner-common:ledger-runner-common
+  type: jar-scala
 - target: //ledger/test-common:dar-files-1.14-lib
   type: jar-scala
 - target: //ledger/test-common:dar-files-1.dev-lib
